+++ spec/Pim/Bundle/CatalogBundle/Manager/GroupManagerSpec.php
@@ -35,6 +35,45 @@ class GroupManagerSpec extends ObjectBehavior
         );
     }
 
+    function it_is_a_saver()
+    {
+        $this->shouldHaveType('Pim\Component\Resource\Model\SaverInterface');
+    }
+    function it_is_a_remover()
+    {
+        $this->shouldHaveType('Pim\Component\Resource\Model\RemoverInterface');
+    }
+
+    function it_throws_exception_when_save_anything_else_than_a_group()
+    {
+        $anythingElse = new \stdClass();
+        $this
+            ->shouldThrow(
+                new \InvalidArgumentException(
+                    sprintf(
+                        'Expects a "Pim\Bundle\CatalogBundle\Entity\Group", "%s" provided.',
+                        get_class($anythingElse)
+                    )
+                )
+            )
+            ->during('save', [$anythingElse]);
+    }
+
+    function it_throws_exception_when_remove_anything_else_than_a_group()
+    {
+        $anythingElse = new \stdClass();
+        $this
+            ->shouldThrow(
+                new \InvalidArgumentException(
+                    sprintf(
+                        'Expects a "Pim\Bundle\CatalogBundle\Entity\Group", "%s" provided.',
+                        get_class($anythingElse)
+                    )
+                )
+            )
+            ->during('remove', [$anythingElse]);
+    }
+
     function it_dispatches_an_event_when_removing_a_group(
         $eventDispatcher,
         $registry,