+++ spec/Pim/Bundle/EnrichBundle/Manager/SequentialEditManagerSpec.php
@@ -91,4 +91,44 @@ class SequentialEditManagerSpec extends ObjectBehavior
 
         $this->findWrap($sequentialEdit, $product);
     }
+
+    function it_is_a_saver()
+    {
+        $this->shouldImplement('Pim\Component\Resource\Model\SaverInterface');
+    }
+
+    function it_is_a_remover()
+    {
+        $this->shouldImplement('Pim\Component\Resource\Model\RemoverInterface');
+    }
+
+    function it_throws_exception_when_save_anything_else_than_a_sequential_edit()
+    {
+        $anythingElse = new \stdClass();
+        $this
+            ->shouldThrow(
+                new \InvalidArgumentException(
+                    sprintf(
+                        'Expects a Pim\Bundle\EnrichBundle\Entity\SequentialEdit, "%s" provided',
+                        get_class($anythingElse)
+                    )
+                )
+            )
+            ->duringSave($anythingElse);
+    }
+
+    function it_throws_exception_when_remove_anything_else_than_a_sequential_edit()
+    {
+        $anythingElse = new \stdClass();
+        $this
+            ->shouldThrow(
+                new \InvalidArgumentException(
+                    sprintf(
+                        'Expects a Pim\Bundle\EnrichBundle\Entity\SequentialEdit, "%s" provided',
+                        get_class($anythingElse)
+                    )
+                )
+            )
+            ->duringRemove($anythingElse);
+    }
 }