+++ src/Pim/Bundle/EnrichBundle/Controller/AttributeController.php
@@ -211,14 +211,15 @@ class AttributeController extends AbstractDoctrineController
         $data = $request->request->all();
 
         if (!empty($data)) {
+            $attributes = [];
             foreach ($data as $id => $sort) {
                 $attribute = $this->getRepository($this->attributeManager->getAttributeClass())->find((int) $id);
                 if ($attribute) {
                     $attribute->setSortOrder((int) $sort);
-                    $this->persist($attribute, false);
+                    $attributes[] = $attribute;
                 }
             }
-            $this->getManagerForClass($this->attributeManager->getAttributeClass())->flush();
+            $this->attributeManager->saveAll($attributes);
 
             return new Response(1);
         }
@@ -258,7 +259,7 @@ class AttributeController extends AbstractDoctrineController
         if ($request->isMethod('POST')) {
             $form->submit($request);
             if ($form->isValid()) {
-                $this->persist($option);
+                $this->optionManager->save($option);
                 $response = array(
                     'status' => 1,
                     'option' => array(